Added the .job_status() API method.
It takes a project name & job id and returns the job status for a single
job; the result can be one of the following: 'running', 'finished' or 'pending'.

Unlike the other API methods, this method does not have a 1 to 1 mapping with
an actual Scrapyd endpoint and as such, is derived. It utilises Scrapyd's
list_jobs endpoint to receive the status for a single job, returning False
if no status for that id could be found.