Cancel job method now returns the previous state
NB: Breaking change

Prior to this commit, the method simply returned True if the "prevstate"
(previous state) of the cancelled spider was reported as "running" by scrapyd.
While correct, the spider could also have been in the "pending" state when
cancelled and in this instance our wrapper would have reported False. This
also went counter to what the docs suggested.

The major version number will be bumped as a result of this change.

Thanks to @BruceDone for raising and fixing the issue as part of #6.